refactor: move evaluators into separate file
